added VariantSelectorEntity and ActualProductEntity to VariantEntity as an Embedded field

+++ app/src/main/java/com/persia/test/data/network/services/persiaatlas/responses/variant/ActualProduct.kt
@@ -3,6 +3,6 @@ package com.persia.test.data.network.services.persiaatlas.responses.variant
 data class ActualProduct(
     val id: Long,
     val title: String,
-    val price_step: Int,
+    val price_step: Long,
     val brand: Brand,
 )
